Reorganize hooks
