+++ src/components/Epub/StatefulReader.tsx
@@ -49,7 +49,7 @@ import { StatefulReaderFooter } from "../StatefulReaderFooter";
 import { usePreferences } from "@/preferences/hooks/usePreferences";
 import { useSettingsComponentStatus } from "@/components/Settings/hooks/useSettingsComponentStatus";
 import { useEpubSettingsCache } from '@/core/Hooks/Epub/useEpubSettingsCache';
-import { useEpubReaderInit } from '@/core/Hooks/useReaderInit';
+import { useEpubReaderInit } from './Hooks/useReaderInit';
 import { useEpubNavigator } from "@/core/Hooks/Epub/useEpubNavigator";
 import { useFullscreen } from "@/core/Hooks/useFullscreen";
 import { usePrevious } from "@/core/Hooks/usePrevious";
